feat: simplified logic of retrieval nodes
